# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost:63531",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/products" : {
      "post" : {
        "tags" : [ "product-controller" ],
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateProductRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/revive" : {
      "post" : {
        "tags" : [ "product-controller" ],
        "operationId" : "revive_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "If-Match",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReviveRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/retire" : {
      "post" : {
        "tags" : [ "product-controller" ],
        "operationId" : "retire",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "If-Match",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RetireRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/publish" : {
      "post" : {
        "tags" : [ "product-controller" ],
        "operationId" : "publish",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PublishClaimRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/merge/{survivorId}" : {
      "post" : {
        "tags" : [ "product-controller" ],
        "operationId" : "merge_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "survivorId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AcceptedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/gtins" : {
      "post" : {
        "tags" : [ "product-controller" ],
        "operationId" : "bindGtin",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GtinBindRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/classify" : {
      "post" : {
        "tags" : [ "product-controller" ],
        "operationId" : "classify",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ClassifyRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/archive" : {
      "post" : {
        "tags" : [ "product-controller" ],
        "operationId" : "archive",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "If-Match",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/activate" : {
      "post" : {
        "tags" : [ "product-controller" ],
        "operationId" : "activate",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "If-Match",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/attributes" : {
      "post" : {
        "tags" : [ "attribute-controller" ],
        "operationId" : "create_2",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateAttributeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IdResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/products/{id}" : {
      "get" : {
        "tags" : [ "product-controller" ],
        "operationId" : "get",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "product-controller" ],
        "operationId" : "patch",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "If-Match",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PatchProductRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/attributes/{key}" : {
      "get" : {
        "tags" : [ "attribute-controller" ],
        "operationId" : "getAttribute",
        "parameters" : [ {
          "name" : "key",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AttributeResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "IdResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "BundleComponentDto" : {
        "type" : "object",
        "properties" : {
          "componentProductId" : {
            "type" : "string"
          },
          "qty" : {
            "type" : "integer",
            "format" : "int32"
          },
          "verticalIdSnapshot" : {
            "type" : "string"
          }
        }
      },
      "CreateProductRequest" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "productType" : {
            "type" : "string"
          },
          "identityType" : {
            "type" : "string"
          },
          "internalKey" : {
            "type" : "string"
          },
          "gtins" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/GtinDto"
            }
          },
          "brandCode" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "verticalId" : {
            "type" : "string"
          },
          "releaseId" : {
            "type" : "string"
          },
          "classificationStatus" : {
            "type" : "string"
          },
          "attributes" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "evidenceRefs" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "bundleContents" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/BundleComponentDto"
            }
          }
        }
      },
      "GtinDto" : {
        "type" : "object",
        "properties" : {
          "value" : {
            "type" : "string"
          },
          "market" : {
            "type" : "string"
          }
        }
      },
      "ProductResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "productType" : {
            "type" : "string"
          },
          "lifecycle" : {
            "type" : "string"
          },
          "brandCode" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "classification" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "attributes" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          },
          "taxonomyPath" : {
            "type" : "string"
          }
        }
      },
      "ReviveRequest" : {
        "type" : "object",
        "properties" : {
          "formulationVersion" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "RetireRequest" : {
        "type" : "object",
        "properties" : {
          "reason" : {
            "type" : "string"
          }
        }
      },
      "PublishClaimRequest" : {
        "type" : "object",
        "properties" : {
          "attributeKey" : {
            "type" : "string"
          },
          "evidenceRefs" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "AcceptedResponse" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          }
        }
      },
      "GtinBindRequest" : {
        "type" : "object",
        "properties" : {
          "gtin" : {
            "type" : "string"
          },
          "market" : {
            "type" : "string"
          }
        }
      },
      "ClassifyRequest" : {
        "type" : "object",
        "properties" : {
          "verticalId" : {
            "type" : "string"
          },
          "releaseId" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "confidence" : {
            "type" : "number",
            "format" : "double"
          },
          "evidenceRefs" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "CreateAttributeRequest" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "governance" : {
            "type" : "string"
          },
          "knownValues" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "PatchProductRequest" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string"
          }
        }
      },
...
      "AttributeResponse" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string"
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          },
          "type" : {
            "type" : "string"
          },
          "governance" : {
            "type" : "string"
          },
          "knownValues" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "status" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}